Update _micro_javascript-performance.md
This is a really cool technique I came across a couple weeks ago. The video I added covers it quite nicely. 

In apps with large config objects/data, it is much faster (up to 2x speed up) to use `JSON.parse()` on an object string instead of writing the object as a POJO. This is because the JIT compiler will parse the object as JavaScript, where the `JSON.parse()` method is focused only on parsing objects (naturally). Or in other words, the JIT compiler has to keep track of context, where the `JSON.parse()` method only has a single context (and the JIT compiler sort of _skips over_ the object as a string - which is super fast.)