Replace WebGL workshop in HTML Journeyman (#114)
The current interactive WebGL Workshop option requires node v4 (breaking change in v5+ that was never fixed). Even with v4 it would crash as you tried making the changes they wanted you to make. There might have been a way to get through it, but it took a lot of time to set up and troubleshoot. The new WebGL workshop is perhaps a bit shorter, but still goes through the basics and explains step by step what we're doing and why we're doing it.